refactor: rename and reorganize blog signup service files
